Fix all clippy warnings and formatting (zero errors on -D warnings)
Applied cargo clippy --fix for ~70 mechanical fixes:
- Collapsible if statements (51 instances across 12 files)
- Manual assign operations → +=, -=
- Unnecessary Ok() wrapping
- match replaceable with ?
- map_or simplifiable
- Derivable impls
- Redundant closures
- Float precision
- assert_eq with literal bool

Manual fixes:
- Delete OrderedCoeff trait, impl, and 3 tests (unused)
- Delete orphaned doc comments in evalf.rs and solve.rs
- Fix doc comment style in units/calculus.rs (/// → //!)
- Fix doc list continuation in risch/integrate.rs
- Fix unwrap-after-is_ok in matrix.rs (use if-let instead)
- Fix needless_range_loop in evalf.rs (use iter().enumerate())
- Allow dead_code on WIP risch/tower.rs module (tested but not yet in production)
- Allow dead_code on WIP risch structs (LogTerm, RischResult, TowerHermiteResult, etc.)
- Allow dead_code on poly/roots.rs functions (rootof_eval_f64, bigfloat_to_f64)
- Allow dead_code on ratfn numer()/denom() (public API for future use)
- Allow type_complexity on units/calculus.rs (7-dimensional type arithmetic)
- Allow needless_range_loop on rde.rs matrix operations
- Allow non_snake_case on pid_controller example (Kp, Ki, Kd are standard notation)

Result: cargo clippy --all-targets -- -D warnings passes with zero errors
        cargo fmt --all -- --check passes with zero diffs
        1882 lib tests pass

+++ tests/proptest_bounded.rs
@@ -206,7 +206,7 @@ fn exhaustive_factor_expand_roundtrip() {
 fn exhaustive_from_conversions() {
     let ctx = Context::new();
     for n in -50i64..=50 {
-        let ex = ctx.int(n as i64);
+        let ex = ctx.int(n);
         assert_eq!(format!("{ex}"), n.to_string(), "From<i64>({n})");
     }
     for n in 0u32..=100 {
